% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_sha1.R
\name{\%<c-\%}
\alias{\%<c-\%}
\title{Assign read in data with creating/checking SHA1 sum}
\usage{
x \%<c-\% value
}
\arguments{
\item{x}{unquoted expression naming variable to create}

\item{value}{unquoted expression to evaluate the first time \code{name} is
accessed}
}
\description{
This infix function creates a normal assignment as side effect,
but also checks the read in data against a previously cached SHA1
sum (if available), or creates the SHA1 sum in the first
place. When SHA1 sums differ, a warning will be issued.
}
\examples{
\dontrun{
x \%<c-\% read.csv("some_file.csv")
}
}
