% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freeze.R
\name{\%<f-\%}
\alias{\%<f-\%}
\title{Infix variant of \code{freeze}}
\usage{
x \%<f-\% value
}
\arguments{
\item{x}{unquoted expression naming variable to create}

\item{value}{unquoted expression to evaluate the first time
\code{name} is accessed}
}
\description{
Does the same as \code{freeze}, but as infix operator.
}
