% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_cache.R
\name{empty_cache}
\alias{empty_cache}
\title{Empty cache}
\usage{
empty_cache()
}
\value{
Nothing, invoked for side effects.
}
\description{
This function removes all objects in the cache, w.r.t. to the current working
directory.
}
\examples{
# empty_cache()
}
