% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forget.R
\name{forget}
\alias{forget}
\title{Forget cached object}
\usage{
forget(name)
}
\arguments{
\item{name}{unquote expression naming object to remove}
}
\value{
nothing, invoked for side effects
}
\description{
Forget cached object
}
