% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fridge_function.R
\name{fridge}
\alias{fridge}
\title{Fridge}
\usage{
fridge(X, y, x0, plug.in = c("OLS", "RLOOCV"), plot.curve = TRUE)
}
\arguments{
\item{X}{Matrix with covariates of dimension n x p}

\item{y}{Vector of outcomes of length n}

\item{x0}{Vector of focus covariate of length p}

\item{plug.in}{Character specifying the plug.in estimate, possibly 'OLS' (ordinary least squares) and 'RLO0CV' (ridge with leave-one-out cross-validation). 
The OLS option can only be used for p smaller than n.}

\item{plot.curve}{Logical. Should the MSE curve and minimum value be ploted?}
}
\value{
Result, a list with the following 
\item{fridge.tuning}{Numerical, the fridge tuning parameter} 
\item{loocv.tuning}{Numerical, the leave-one-out cross-validation tuning parameter}
\item{fridge.pred}{Numerical, the fridge prediction} 
\item{loocv.pred}{Numerical, the leave-one-out cross-validation prediction}
}
\description{
Function for calculating the fridge-ridge tuning parameter and prediction, with plot of MSE curve
}
\details{
The minimization routine is carried out with several different start values ranging from 1 to 10^7. If plot.curve = TRUE, the global minimum
is ploted.
}
\examples{
fridge()
}
\keyword{Focused}
\keyword{tuning}
