% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{request_all}
\alias{request_all}
\title{Request objects by pattern from session or cache}
\usage{
request_all(.pattern)
}
\arguments{
\item{.pattern}{a regular expression as string}
}
\value{
Nothing, invoked for side effects (requesting objects)
}
\description{
This functions works like \code{request}, but takes a
  regular expression to match against objects in workspace and
  cache.
}
