% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fridge_functions.R
\name{tuning.fridge.ols.sim}
\alias{tuning.fridge.ols.sim}
\title{Simplified fridge-OLS}
\usage{
tuning.fridge.ols.sim(svd, sigma2_hat, lambda, x0)
}
\arguments{
\item{svd}{SVD of the n x p data matrix}

\item{sigma2_hat}{Plug-in estimate of the error variance}

\item{lambda}{Ridge tuning parameter}

\item{x0}{Focus covariate}
}
\description{
Function for fridge-OLS error
}
\details{
Can only be used in the low-dimensional case, for n < p

tuning.fridge.ols.sim
}
