% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fridge_functions.R
\name{tuning.fridge.ridge.sim}
\alias{tuning.fridge.ridge.sim}
\title{Simplified fridge-ridge}
\usage{
tuning.fridge.ridge.sim(svd, sigma2_hat, lambda, x0, cv.tuning)
}
\arguments{
\item{svd}{SVD of the n x p data matrix}

\item{sigma2_hat}{Plug-in estimate of the error variance}

\item{lambda}{Ridge tuning parameter}

\item{x0}{Focus covariate}

\item{init.tuning}{Initial ridge tuning parameter}
}
\description{
Function for fridge-ridge error
}
\details{
The function gives error for fridge-ridge estimator

tuning.fridge.ridge.sim
}
