% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_packages.R
\name{load_packages}
\alias{load_packages}
\title{Load packages}
\usage{
load_packages(packages)
}
\arguments{
\item{packages}{a character vector holding the names of packages
to load}
}
\value{
nothing, invoked for side effects (package loading)
}
\description{
Load a number of packages without needing a new line of
"library(...)" for each one.
}
