% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{request}
\alias{request}
\title{Request object from session or cache}
\usage{
request(.name)
}
\arguments{
\item{.name}{String giving the name of the requested object}
}
\value{
Nothing, invoked for side effects (requesting an object)
}
\description{
This function is a wrapper around \link{\code{thaw}}, that checks if an
object is currently defined in the workspace. If not, it tries to load the
object from the project's cache.
}
