% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lib.R
\name{thaw}
\alias{thaw}
\title{Load cached objects from /cache directory of project tree}
\usage{
thaw(str)
}
\arguments{
\item{str}{Name of object saved rda file}
}
\value{
nothing, invoked for side effects
}
\description{
Load cached objects from /cache directory of project tree
}
\details{
To save time and typing, this function will load rda files from /cache
directory of project tree by name.
}
\examples{
\dontrun{thaw("someobject")}
}
\author{
Christian Zang
}
