% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fridge_function.R
\name{tuning.fridge.ols.sim}
\alias{tuning.fridge.ols.sim}
\title{Simplified fridge-OLS}
\usage{
tuning.fridge.ols.sim(lambda, svd, sigma2_hat, x0, y)
}
\arguments{
\item{lambda}{Ridge tuning parameter}

\item{svd}{SVD of the n x p data matrix}

\item{sigma2_hat}{Plug-in estimate of the error variance}

\item{x0}{Focus covariate}
}
\description{
Function for fridge-OLS error
}
\details{
Can only be used in the low-dimensional case, for n < p
}
\examples{
tuning.fridge.ols.sim
}
\keyword{Focused}
\keyword{tuning}
