% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fridge_function.R
\name{tuning.fridge.ridge.sim}
\alias{tuning.fridge.ridge.sim}
\title{Simplified fridge-ridge}
\usage{
tuning.fridge.ridge.sim(lambda, svd, y, sigma2_hat, x0, cv.tuning)
}
\arguments{
\item{lambda}{Ridge tuning parameter}

\item{svd}{SVD of the n x p data matrix}

\item{sigma2_hat}{Plug-in estimate of the error variance}

\item{x0}{Focus covariate}

\item{init.tuning}{Initial ridge tuning parameter}
}
\description{
Function for fridge-ridge error
}
\details{
The function gives error for fridge-ridge estimator
}
\examples{
tuning.fridge.ridge.sim
}
\keyword{Focused}
\keyword{tuning}
