% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fridge_functions.R
\name{tuning.loocv}
\alias{tuning.loocv}
\title{Cross-validation}
\usage{
tuning.loocv(svd, n, lambda)
}
\arguments{
\item{svd}{SVD of the n x p data matrix}

\item{n}{Number of observations}

\item{lambda}{Ridge tuning parameter}
}
\description{
Function for the ridge regression leave-one-out cross-validation error
}
\details{
The function gives the explitic expression of the LOOCV error for ridge regression

tuning.loocv
}
